chore: renamed some packages and folders and methods to better fit golang naming conventions
