+++ cmd/app/main.go
@@ -1,18 +1,19 @@
 package main
 
 import (
-	controller "api/internal/app/http"
-	"api/internal/app/router"
+	"api/internal/app/controller"
+	"api/internal/app/routing"
 )
 
 func main() {
 	// Create a new router
-	router := router.NewRouter()
+	router := routing.New()
 	// Add routes to the router
 	router.Handle("/api/v1/auth/login", controller.Test())
 
 	// Start the server on port 8080
-	router.Serve(8080)
+
+	router.Serve(8080, routing.ServeOptions{Message: "http://localhost:8080",})
 
 }
